+++ fast_healthchecks/checks/postgresql/asyncpg.py
@@ -69,7 +69,7 @@ class PostgreSQLAsyncPGHealthCheck(BasePostgreSQLHealthCheck[HealthCheckResult])
         *,
         config: PostgresAsyncPGConfig | None = None,
         name: str = "PostgreSQL",
-        **kwargs: Any,  # noqa: ANN401
+        **kwargs: Any,  # ruff: ignore[any-type]
     ) -> None:
         """Initialize the PostgreSQLAsyncPGHealthCheck.
 
@@ -86,7 +86,7 @@ class PostgreSQLAsyncPGHealthCheck(BasePostgreSQLHealthCheck[HealthCheckResult])
     @classmethod
     def _from_parsed_dsn(
         cls,
-        parsed: "PostgresParseDsnResult",  # noqa: UP037
+        parsed: "PostgresParseDsnResult",  # ruff: ignore[quoted-annotation]
         *,
         name: str = "PostgreSQL",
         timeout: float = DEFAULT_HC_TIMEOUT,