+++ fast_healthchecks/checks/rabbitmq.py
@@ -89,7 +89,7 @@ class RabbitMQHealthCheck(
         config: RabbitMQConfig | None = None,
         name: str = "RabbitMQ",
         close_client_fn: Callable[[AbstractRobustConnection], Awaitable[None]] = _close_rabbitmq_client,
-        **kwargs: Any,  # noqa: ANN401
+        **kwargs: Any,  # ruff: ignore[any-type]
     ) -> None:
         """Initialize the RabbitMQHealthCheck.